# ARROW-8831: [Rust] change simd_compare_op in comparison kernel to use bitmask SIMD operation to significantly improve performance