--- a/rust/arrow/src/compute/kernels/comparison.rs
+++ b/rust/arrow/src/compute/kernels/comparison.rs
@@ -198,6 +198,10 @@ pub fn gt_eq_utf8(left: &StringArray, right: &StringArray) -> Result<BooleanArra
     compare_op!(left, right, |a, b| a >= b)
 }
 
+use crate::buffer::MutableBuffer;
+use std::mem;
+use std::io::Write;
+
 /// Helper function to perform boolean lambda function on values from two arrays using
 /// SIMD.
 #[cfg(all(any(target_arch = "x86", target_arch = "x86_64"), feature = "simd"))]
@@ -225,7 +229,7 @@ where
     )?;
 
     let lanes = T::lanes();
-    let mut result = BooleanBufferBuilder::new(len);
+    let mut result = MutableBuffer::new(left.len() * mem::size_of::<bool>());
 
     let rem = len % lanes;
 
@@ -233,18 +237,15 @@ where
         let simd_left = T::load(left.value_slice(i, lanes));
         let simd_right = T::load(right.value_slice(i, lanes));
         let simd_result = op(simd_left, simd_right);
-        for i in 0..lanes {
-            result.append(T::mask_get(&simd_result, i))?;
+        T::bitmask(&simd_result, |b| { result.write(b); });
         }
-    }
 
     if rem > 0 {
         let simd_left = T::load(left.value_slice(len - rem, lanes));
         let simd_right = T::load(right.value_slice(len - rem, lanes));
         let simd_result = op(simd_left, simd_right);
-        for i in 0..rem {
-            result.append(T::mask_get(&simd_result, i))?;
-        }
+        let rem_buffer_size = (rem as f32 / 8f32).ceil() as usize;
+        T::bitmask(&simd_result, |b| { result.write(&b[0..rem_buffer_size]); });
     }
 
     let data = ArrayData::new(
@@ -253,7 +254,7 @@ where
         None,
         null_bit_buffer,
         left.offset(),
-        vec![result.finish()],
+        vec![result.freeze()],
         vec![],
     );
     Ok(PrimitiveArray::<BooleanType>::from(Arc::new(data)))