--- a/rust/arrow/src/datatypes.rs
+++ b/rust/arrow/src/datatypes.rs
@@ -521,7 +521,9 @@ where
     fn mask_get(mask: &Self::SimdMask, idx: usize) -> bool;
 
     /// Gets the bitmask for a SimdMask as a byte slice and passes it to the closure used as the action parameter
-    fn bitmask<T>(mask: &Self::SimdMask, action: T) where T: FnMut(&[u8]);
+    fn bitmask<T>(mask: &Self::SimdMask, action: T)
+    where
+        T: FnMut(&[u8]);
 
     /// Sets the value of a single lane of a SIMD mask
     fn mask_set(mask: Self::SimdMask, idx: usize, value: bool) -> Self::SimdMask;
@@ -595,7 +597,10 @@ macro_rules! make_numeric_type {
                 unsafe { mask.extract_unchecked(idx) }
             }
 
-            fn bitmask<T>(mask: &Self::SimdMask, mut action: T) where T: FnMut(&[u8]) {
+            fn bitmask<T>(mask: &Self::SimdMask, mut action: T)
+            where
+                T: FnMut(&[u8]),
+            {
                 action(mask.bitmask().to_byte_slice());
             }